Fix crash opening the movement menu (64-bit pointer truncation)

+++ Tactical/Interface.cpp
@@ -843,7 +843,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	SetButtonFastHelpText( iActionIcons[ RUN_ICON ], pTacticalPopupButtonStrings[ RUN_ICON ] );
 	//SetButtonSavedRect( iActionIcons[ RUN_ICON ] );
-	ButtonList[ iActionIcons[ RUN_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ RUN_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( pSoldier->MercInWater( ) || ( pSoldier->flags.uiStatusFlags & SOLDIER_ROBOT ) )
 	{
@@ -869,7 +869,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 		SetButtonFastHelpText( iActionIcons[ WALK_ICON ], pTacticalPopupButtonStrings[ WALK_ICON ] );
 	}
 
-	ButtonList[ iActionIcons[ WALK_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ WALK_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( pSoldier->flags.uiStatusFlags & SOLDIER_ROBOT )
 	{
@@ -890,7 +890,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	SetButtonFastHelpText( iActionIcons[ SNEAK_ICON ], pTacticalPopupButtonStrings[ SNEAK_ICON ] );
 	//SetButtonSavedRect( iActionIcons[ SNEAK_ICON ] );
-	ButtonList[ iActionIcons[ SNEAK_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ SNEAK_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	// Check if this is a valid stance, diable if not!
 	if ( !IsValidStance( pSoldier, ANIM_CROUCH ) )
@@ -908,7 +908,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	SetButtonFastHelpText( iActionIcons[ CRAWL_ICON ], pTacticalPopupButtonStrings[ CRAWL_ICON ] );
 	//SetButtonSavedRect( iActionIcons[ CRAWL_ICON ] );
-	ButtonList[ iActionIcons[ CRAWL_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ CRAWL_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	// Check if this is a valid stance, diable if not!
 	if ( !IsValidStance( pSoldier, ANIM_PRONE ) )
@@ -927,7 +927,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	SetButtonFastHelpText( iActionIcons[ LOOK_ICON ], TacticalStr[ LOOK_CURSOR_POPUPTEXT ] );
 	//SetButtonSavedRect( iActionIcons[ LOOK_ICON ] );
-	ButtonList[ iActionIcons[ LOOK_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ LOOK_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( pSoldier->flags.uiStatusFlags & SOLDIER_VEHICLE )
 	{
@@ -1033,7 +1033,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	//SetButtonSavedRect( iActionIcons[ ACTIONC_ICON ] );
 	SetButtonFastHelpText( iActionIcons[ ACTIONC_ICON ], zActionString );
-	ButtonList[ iActionIcons[ ACTIONC_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ ACTIONC_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( fDisableAction )
 	{
@@ -1051,7 +1051,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	//SetButtonSavedRect( iActionIcons[ TALK_ICON ] );
 	SetButtonFastHelpText( iActionIcons[ TALK_ICON ], pTacticalPopupButtonStrings[ TALK_ICON ] );
-	ButtonList[ iActionIcons[ TALK_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ TALK_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( AM_AN_EPC( pSoldier ) || ( pSoldier->flags.uiStatusFlags & SOLDIER_VEHICLE ) )
 	{
@@ -1069,7 +1069,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	//SetButtonSavedRect( iActionIcons[ HAND_ICON ] );
 	SetButtonFastHelpText( iActionIcons[ HAND_ICON ], pTacticalPopupButtonStrings[ HAND_ICON ] );
-	ButtonList[ iActionIcons[ HAND_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ HAND_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	if ( AM_AN_EPC( pSoldier ) || ( pSoldier->flags.uiStatusFlags & SOLDIER_VEHICLE ) )
 	{
@@ -1086,7 +1086,7 @@ void PopupMovementMenu( UI_EVENT *pUIEvent )
 	}
 	//SetButtonSavedRect( iActionIcons[ CANCEL_ICON ] );
 	SetButtonFastHelpText( iActionIcons[ CANCEL_ICON ], pTacticalPopupButtonStrings[ CANCEL_ICON ] );
-	ButtonList[ iActionIcons[ CANCEL_ICON ] ]->UserData[0] = (UINT32)(uintptr_t)pUIEvent;
+	ButtonList[ iActionIcons[ CANCEL_ICON ] ]->UserData[0] = (uintptr_t)pUIEvent;
 
 	//LockTacticalInterface( );
